benchmarks/competitive/requirements.txt
# Competitive Benchmarking Dependencies

# Vector database clients
pinecone-client==3.0.0
qdrant-client==1.7.0
# d-vecdb client (local)
# vectordb-client  # Install from ../python-client

# Data generation and manipulation
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Benchmarking and profiling
pytest==7.4.0
pytest-benchmark==4.0.0
memory-profiler==0.61.0
psutil==5.9.5

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Reporting
tabulate==0.9.0
jinja2==3.1.2

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
pyyaml==6.0.1
